# lib/api-zod/src/generated/types/donation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnnSetu API - Food donation platform connecting donors with NGOs and volunteers
 * OpenAPI spec version: 0.1.0
 */
import type { DonationFoodType } from './donationFoodType';
import type { DonationStatus } from './donationStatus';
import type { UserProfile } from './userProfile';

export interface Donation {
  id: number;
  donorId: number;
  donor?: UserProfile;
  foodName: string;
  foodType: DonationFoodType;
  quantityPlates: number;
  /** @nullable */
  estimatedServings?: number | null;
  /** @nullable */
  preparedAt?: Date | null;
  pickupDeadline?: Date;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  status: DonationStatus;
  /** @nullable */
  claimedByUserId?: number | null;
  claimedBy?: UserProfile | null;
  createdAt: Date;
  updatedAt?: Date;
}