# lib/api-zod/src/generated/types/donationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnnSetu API - Food donation platform connecting donors with NGOs and volunteers
 * OpenAPI spec version: 0.1.0
 */
import type { DonationInputFoodType } from './donationInputFoodType';

export interface DonationInput {
  foodName: string;
  foodType: DonationInputFoodType;
  /** @minimum 1 */
  quantityPlates: number;
  estimatedServings?: number;
  preparedAt?: Date;
  pickupDeadline: Date;
  description?: string;
  imageUrl?: string;
  address?: string;
  lat?: number;
  lng?: number;
}