# lib/api-zod/src/generated/types/donationUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnnSetu API - Food donation platform connecting donors with NGOs and volunteers
 * OpenAPI spec version: 0.1.0
 */
import type { DonationUpdateFoodType } from './donationUpdateFoodType';
import type { DonationUpdateStatus } from './donationUpdateStatus';

export interface DonationUpdate {
  foodName?: string;
  foodType?: DonationUpdateFoodType;
  quantityPlates?: number;
  estimatedServings?: number;
  preparedAt?: Date;
  pickupDeadline?: Date;
  description?: string;
  imageUrl?: string;
  address?: string;
  lat?: number;
  lng?: number;
  status?: DonationUpdateStatus;
}