# lib/api-zod/src/generated/types/ngoStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnnSetu API - Food donation platform connecting donors with NGOs and volunteers
 * OpenAPI spec version: 0.1.0
 */
import type { Claim } from './claim';

export interface NgoStats {
  totalClaims: number;
  completedPickups: number;
  totalPlatesCollected: number;
  activeClaims: number;
  recentClaims?: Claim[];
}