# lib/api-zod/src/generated/types/otpVerifyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnnSetu API - Food donation platform connecting donors with NGOs and volunteers
 * OpenAPI spec version: 0.1.0
 */

export interface OtpVerifyInput {
  otp: string;
}