# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnnSetu API - Food donation platform connecting donors with NGOs and volunteers
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileInputDonorCategory } from './userProfileInputDonorCategory';
import type { UserProfileInputRole } from './userProfileInputRole';

export interface UserProfileInput {
  role: UserProfileInputRole;
  name: string;
  phone: string;
  address?: string;
  city?: string;
  lat?: number;
  lng?: number;
  donorCategory?: UserProfileInputDonorCategory;
  licenseNumber?: string;
  orgName?: string;
  registrationNumber?: string;
  operatingRadiusKm?: number;
  vehicleType?: string;
  availabilityStatus?: string;
}